+++ pyvela/pyvela/spnta.py
@@ -173,6 +173,10 @@ class SPNTA:
     @property
     def param_units(self) -> Iterable[str]:
         return np.array(list(vl.get_free_param_units(self.pulsar.model)))
+    
+    @property
+    def param_prefixes(self) -> Iterable[str]:
+        return np.array(list(vl.get_free_param_prefixes(self.pulsar.model)))
 
     @property
     def scale_factors(self) -> Iterable[float]: